requirements.txt
gevent
pymongo>=3.6
pymysql>=0.8.0
pyyaml>=3.12
pytest
rethinkdb<=2.2
